requirements.txt
matplotlib==2.2.2
numpy==1.15.4
Pillow==5.3.0
torch==1.0
torchvision==0.2.1
grpcio==1.17.0
grpcio-tools==1.17.0
argparse==1.4.0
opencv-python
python-magic
